solid/SRP/solution/requirements.txt
# Pinned requirements for reproducibility
torch==2.4.1
pandas==2.2.3
pyyaml==6.0.2
scikit-learn==1.5.2
pytest==8.3.3

